* generic/tclOOStubLib.c:	ensure use of tcl stubs; include in 	* unix/Makefile.in:		stub lib; disable broken tclOO genstubs
	* generic/tclOO.c:		make tclOO stubs tables 'static const'
	* generic/tclOODecls.h:		and stub table pointers MODULE_SCOPE
	* generic/tclOOIntDecls.h:	(change generated files manually
	* generic/tclOOStubInit.c:	pending genstubs support for tclOO).
	* generic/tclOOStubLib.c:

	* generic/tclOO.c:		fix warnings for 'int<->ptr conversion'
	* generic/tclOOCall.c:		and 'signed vs unsigned comparison'.
	* generic/tclOOMethod.c:

+++ generic/tclOOMethod.c
@@ -915,7 +915,7 @@ ConstructorErrorHandler(
     const char *objectName, *kindName;
     int objectNameLen;
 
-    if (interp->errorLine == 0xDEADBEEF) {
+    if (interp->errorLine == (int) 0xDEADBEEF) {
 	/*
 	 * Horrible hack to deal with certain constructors that must not add
 	 * information to the error trace.
